refactor: create blitzar_metadata_table module from dynamic_dory_commitment_helper_gpu code
